refactor: split out main window panels
